+++ crates/holochain/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Remove the previously deprecated type `CellStatus`.
+
 ## 0.5.0-dev.20
 
 ## 0.5.0-dev.19